Add Money.convert/2
Note that this is a toy library for personal use. If this lib were
intended for others to use or in a production environment, I would
created bang versions for all these operations which is the version that
would raise if usage is incorrect within the domain. The non-bang
versions would return :ok/:error tuples for better control flow.