Add Money.to_string/1
Will add options later to control the formatting.

It works by recursively iterating over the digits in reverse with their
index, while building up an accumulator that has the end result in order.

The digits are reversed because it makes it easier to work
right-to-left when handling delimiters and decimal points.

The index helps to keep track of progress within the remaining digits,
so that we can add delimiters in the right place.

The base cases are
1. Values in the hundredths with no remainder, e.g. 5 -> $0.05
2. Values in the tenths with no remainder, e.g. 50 -> $0.50
3. Values with no remainder -> done recursing, return accumulated list

A non-base case that we need to handle is adding a decimal point while
continuing to recurse. This happens when we have numbers remaining and
two fixed numbers in the accumulator.

The final case recurses over the remainder and depending on the index of
the head digit, may also add a comma as a separator (delimiter).

Example iteration:

123_456 => "£1,234.56"

[{"6", 0}, {"5", 1}, {"4", 2}, {"3", 3}, {"2", 4}, {"1", 5}]  []
[{"5", 1}, {"4", 2}, {"3", 3}, {"2", 4}, {"1", 5}]            ["6"]
[{"4", 2}, {"3", 3}, {"2", 4}, {"1", 5}]                      ["5", "6"]
[{"3", 3}, {"2", 4}, {"1", 5}]                                ["4", ".", "5", "6"]
[{"2", 4}, {"1", 5}]                                          ["3", "4", ".", "5", "6"]
[{"1", 5}]                                                    ["2", "3", "4", ".", "5", "6"]
[]                                                            ["1", ",", "2", "3", "4", ".", "5", "6"]