Fix pattern matching bug
The pattern +0.0 fails to match on 0 and you can get an infinite loop
due to recursion not handling that case.

We're going to move away from floats as money should never be
represented that way, too error-prone. Was only for quick prototyping.
We now have the money module as a dependency that we can switch to.